refac: useScrollTo 리팩토링
